move to a unified newSources
