+++ src/actions/index.js
@@ -18,6 +18,7 @@ import * as projectTextSearch from "./project-text-search";
 import * as quickOpen from "./quick-open";
 import * as sourceTree from "./source-tree";
 import * as loadSourceText from "./sources/loadSourceText";
+import * as newSources from "./sources/newSources";
 import * as debuggee from "./debuggee";
 import * as toolbox from "./toolbox";
 
@@ -36,6 +37,7 @@ export default {
   ...quickOpen,
   ...sourceTree,
   ...loadSourceText,
+  ...newSources,
   ...debuggee,
   ...toolbox
 };
